+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/website/LibraryController.scala
@@ -61,25 +61,7 @@ class LibraryController @Inject() (
     extends UserActions with LibraryAccessActions with ShoeboxServiceController with Logging {
 
   private def getSuggestedSearchesAsJson(libId: Id[Library]): JsValue = {
-    def similar(s1: String, s2: String): Boolean = {
-      val ts1 = s1.split(" ").flatMap(_.sliding(2)).toSet
-      val ts2 = s2.split(" ").flatMap(_.sliding(2)).toSet
-      ts1.intersect(ts2).size * 1.0 / ts1.union(ts2).size > 0.6
-    }
-
-    val LIMIT = 10
-    val top = suggestedSearchCommander.getSuggestedTermsForLibrary(libId, limit = LIMIT * 2, kind = SuggestedSearchTermKind.AUTO)
-    val (terms, weights): (Array[String], Array[Float]) = if (top.terms.size < LIMIT) (Array[String](), Array[Float]()) else {
-      val (tms, ws) = top.terms.toArray.sortBy(-_._2).unzip
-      val taken = mutable.Set[String]()
-      tms.foreach { tm =>
-        val dup = taken.find(x => similar(tm, x))
-        if (dup.isEmpty) taken.add(tm)
-      }
-
-      val (tms2, ws2) = (tms zip ws).filter { case (word, weight) => taken.contains(word) }.take(LIMIT).unzip
-      (tms2.toArray, ws2.toArray)
-    }
+    val (terms, weights) = suggestedSearchCommander.getTopTermsForLibrary(libId, limit = 10)
     Json.obj("terms" -> terms, "weights" -> weights)
   }
